+++ quanmiaolightapp-20240801/ChangeLog.md
@@ -1,3 +1,10 @@
+2024-09-26 Version: 1.3.0
+- Support API GenerateBroadcastNews.
+- Support API RunCommentGeneration.
+- Update API RunVideoAnalysis: add param snapshotInterval.
+- Update API RunVideoAnalysis: add param videoExtraInfo.
+
+
 2024-09-14 Version: 1.2.0
 - Support API ListHotTopicSummaries.
 - Update API RunVideoAnalysis: update response param.